Add build step via Babel, convert to ES module imports
This gives us the benefit of smaller bundle sizes on the front-end - if
you run the `build` script and take a look at any of the files in the
`dist/losant` folder you'll see we're now importing everything directly
from Ramda / Lodash (e.g. `require('ramda/src/compose')`).

The Babel configuration I have for the front-end works well enough when
we're importing functions directly from Ramda or Lodash but doesn't work
on the imports for our custom functions.